requirements.txt
python-telegram-bot>=20.0
openai-whisper>=20231117
google-auth-oauthlib>=1.0.0
google-auth-httplib2>=0.1.0
google-api-python-client>=2.0.0
google-generativeai>=0.3.0
python-dotenv>=1.0.0
ffmpeg-python>=0.2.0 